♻ Refactored the code to split the logic
